🚀 [Deploy] - Choco API 예외처리를 서버에 반영한다 (#27)
* Refactor: 디렉토리 이름 수정

* ✨ [Feature] - Choco API를 구현한다 (#8)

* Refactor: 디렉토리 이름 수정

* Feat: Choco API 생성 완료

* Chore: Redis credential 수정 (#11)

* Chore: CI/CD 진행 시에만 바스티온 호스트의 ingress 22번 포트를 뚫을 수 있도록 한다 (#14)

* Chore: Github actions IP 오류를 해결한다 (#17)

* 🖊️ [Chore] - AWS 접속 시, 발생하는 오류를 해결한다 (#19)

* Chore: Github actions IP 오류를 해결한다

* Chore: Github actions IP 오류를 해결한다

* Chore: CD 진행 중 에러 발생 트러블 슈팅

* Chore/#16 (#22)

* Chore: Github actions IP 오류를 해결한다

* Chore: Github actions IP 오류를 해결한다

* Chore: Redis에서 Session을 사용한다

* Feat: Choco API 예외처리 구현

* Fix: Choco API 예외처리 버그 수정

---------

+++ src/main/java/dgu/choco_express/service/choco/ChocoService.java
@@ -36,17 +36,19 @@ public class ChocoService {
             ChocoCreateRequestDto chocoCreateRequestDto
     ) {
         User currentUser = userRetriever.findById(userId);
-
         Short chocoType = chocoCreateRequestDto.chocoType();
         String chocoNickname = chocoCreateRequestDto.nickname();
         String chocoContents = chocoCreateRequestDto.contents();
         Box currentBox = boxRetriever.findById(boxId);
 
-
         if (chocoType < 1 || chocoType > 6)
             throw CommonException.type(ChocoErrorCode.INVALID_CHOCO_TYPE);
         if (chocoNickname.isEmpty())
             throw CommonException.type(ChocoErrorCode.INVALID_CHOCO_NAME);
+        if (chocoContents.isEmpty())
+            throw CommonException.type(ChocoErrorCode.INVALID_CHOCO_CONTENT);
+        if (currentUser.equals(currentBox.getUser()))
+            throw CommonException.type(ChocoErrorCode.CANT_CHOCO_RECURSIVE);
 
         Choco createdChoco = chocoSaver.saveChoco(
             Choco.from(chocoType, chocoNickname, chocoContents, currentBox)
@@ -91,6 +93,9 @@ public class ChocoService {
         User currentUser = userRetriever.findById(userId);
         Choco choco = chocoRetriever.findById(chocoId);
 
+        if (!currentUser.equals(choco.getBox().getUser()))
+            throw CommonException.type(ChocoErrorCode.CANT_READ_CHOCO);
+
         return ChocoDetailsResponseDto.of(choco);
     }
 
@@ -99,6 +104,9 @@ public class ChocoService {
     ) {
         User currentUser = userRetriever.findById(userId);
         Box currentBox = boxRetriever.findByUser(currentUser);
+
+        if (!currentUser.equals(currentBox.getUser()))
+            throw CommonException.type(ChocoErrorCode.CANT_READ_BOX);
 
         return ChocoPeekResponseDto.builder()
                 .count(chocoRetriever.findChocoCountByBox(currentBox))
@@ -111,6 +119,9 @@ public class ChocoService {
     ) {
         User currentUser = userRetriever.findById(userId);
         Choco choco = chocoRetriever.findById(chocoId);
+
+        if(!currentUser.equals(choco.getBox().getUser()))
+            throw CommonException.type(ChocoErrorCode.CANT_DELETE_CHOCO);
 
         chocoRemover.deleteChoco(choco);
